Cleaning...
I added icons to items in the drawer menu.
I changed the panel_index intialization from main activity to the
partilucar fragment (as things should had been).
I added a checkbox in Network Configuration activity instead of using
other_panel_index.
Next : to work on the local or Internet activity (now it's not the
desired one, just a copy of another activity)