Allowing ports and intermediate relay server IP assignment
I added a section in Network Configuration activity/class for manually
assigning the values of the intermediate IP (which usually should be
hidden from the user), and the ports as well.
I also made a password functionality.
Note that the changes in this section only affect the mobile shared
preferences of the particular panel and I think I won't configure the
panel through it, I mean it's still static in the arduino code.
Later of course, it's interesting to configure the panel as well through
it.